% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa.R
\docType{package}
\name{wdpa}
\alias{wdpa}
\alias{wdpa-package}
\title{wdpa: Shape files of the WDPA project.}
\description{
The wdpa package contains the shape files of the WDPA (www.wdpa.org) project.
So far only the 26 IUCN category-II protected areas of Vietnam are included.
They are in the "vietnam" SpatialPolygonsDataFrame object.
}

